Revert "feat(form): add react-window virtualized list for model field"
