+++ packages/form/package.json
@@ -39,11 +39,10 @@
     "@mui/x-date-pickers": "^6.9.0",
     "@mui/x-date-pickers-pro": "^6.9.0",
     "downloadjs": "^1.4.7",
+    "react-hook-form": "^7.30.0",
     "lodash": "^4.17.21",
     "quill": ">=1.3.7",
-    "react-hook-form": "^7.30.0",
     "react-quilljs": ">=1.2.17",
-    "react-window": "^1.8.9",
     "tslib": "^2.4.0",
     "typescript": "^4.6.3"
   },
