--- packages/form/src/form/utils/getIsCreateOption.ts
@@ -1,6 +1,0 @@
-import get from 'lodash/get'
-
-export const getIsCreateOption = ({ option, pk }) => {
-  if (!get(option, pk)) return
-  return option && pk && get(option, pk, '').toString().startsWith('Add "')
-}
