--- packages/form/src/types.ts
@@ -1,4 +1,0 @@
-export interface DataItem {
-  [key: string]: unknown
-  id?: string | number
-}
